requirement.txt
Flask==3.0.3
Pillow==10.3.0
python-dotenv==1.0.1
langchain==0.1.16
langchain-openai==0.1.6
openai==1.30.1
